+++ .gitignore
@@ -5,3 +5,5 @@ coverage
 *.min.js
 *.min.css
 *.map
+js/block-editor.js
+js/classic-editor.js
